refactored stuff, added new placeholders
